+++ app/initial_state.js
@@ -139,7 +139,6 @@ const state = {
         root: {
             deepLinkURL: '',
             hydrationComplete: false,
-            purge: false,
         },
         selectServer: {
             serverUrl: Config.DefaultServerUrl,
