+++ app/store/middleware.test.js
@@ -14,7 +14,7 @@ import {
 } from 'app/store/middleware';
 
 describe('messageRetention', () => {
-    const messageRetention = middlewares[0];
+    const messageRetention = middlewares()[0];
 
     describe('should chain the same incoming action type', () => {
         const actions = [
